refactor: 패키징 적용
