+++ app/src/main/java/woowacourse/movie/domain/TicketCount.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.domain
 
 @JvmInline
 value class TicketCount(val value: Int = MIN_TICKET_COUNT) {
