+++ app/src/main/java/woowacourse/movie/ui/BookingActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.ui
 
 import android.content.Context
 import android.content.Intent
@@ -7,6 +7,11 @@ import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
+import woowacourse.movie.R
+import woowacourse.movie.domain.Movie
+import woowacourse.movie.domain.MovieData
+import woowacourse.movie.domain.TicketCount
+import woowacourse.movie.formatScreenDate
 
 class BookingActivity : AppCompatActivity() {
     var ticketCount = TicketCount()
