+++ app/src/main/java/woowacourse/movie/ui/MainActivity.kt
@@ -1,8 +1,10 @@
-package woowacourse.movie
+package woowacourse.movie.ui
 
 import android.os.Bundle
 import android.widget.ListView
 import androidx.appcompat.app.AppCompatActivity
+import woowacourse.movie.R
+import woowacourse.movie.domain.MovieData
 
 class MainActivity : AppCompatActivity() {
     private val movieAdapter by lazy { MovieAdapter(this) { clickBook(it) } }
