+++ app/src/main/res/layout/activity_completed.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CompletedActivity">
+    tools:context=".ui.CompletedActivity">
 
     <TextView
         android:id="@+id/textCompletedDescription"
