+++ app/src/test/java/woowacourse/movie/TicketCountTest.kt
@@ -2,6 +2,7 @@ package woowacourse.movie
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
+import woowacourse.movie.domain.TicketCount
 
 internal class TicketCountTest {
     @Test
